simplify top level makefile
The top level Makefile copied too much code from the
freedom-e-sdk/Makefile, making it fragile as the sdk changed.

Strip off a lot of that dependence, having the Makefile call into the
SDK Makefile for most of the work.  It still has some dependence on
the SDK Makefile, but greatly reduced.

Signed-off-by: Curt Brune <curt@brune.net>